fix(post/put/delete products): added axios config for protected routes
